+++ R/pcode_to_name.R
@@ -1,64 +1,24 @@
 #' Parameter code to characteristic name
 #'
-#' This function is useful to fine what characteristic name, result sample
-#' fraction, unit code, and other parameters are mapped with USGS parameter
-#' codes. This information is useful for converting workflows from a more
-#' traditional NWIS water quality retrieval to a Water Quality Portal retrieval.
+#' This function no longer works because the underlying service
+#' has been removed. Similar functionality can be found with
+#' `check_waterdata_sample_params`.
 #'
 #' @export
 #' @param parameterCd character that contains the code for a character vector
-#' of 5-digit parameter codes. Default is "all" which will return a complete
-#' list of parameter codes that have been mapped to a characteristic name.
-#' @return a data frame with columns  "parm_cd", "description",
-#' "characteristicname", "measureunitcode", "resultsamplefraction",
-#' "resulttemperaturebasis", "resultstatisticalbasis",  "resulttimebasis",
-#' "resultweightbasis", "resultparticlesizebasis", "last_rev_dt"
+#' of 5-digit parameter codes.
 #' @examples
 #' pcodes <- c("00070", "00075", "00430", "52642")
 #' \donttest{
-#'
-#' all <- pcode_to_name()
-#' some <- pcode_to_name(pcodes)
+#' all_new <- read_waterdata_parameter_codes(parameter_code = pcodes)
+#' ref_list <- check_waterdata_sample_params("reference-list")
+#' ref_list_sm <- ref_list[ref_list$parm_cd %in% pcodes, ]
 #'
 #' }
 pcode_to_name <- function(parameterCd = "all") {
-  parameterCd.orig <- parameterCd
-  parameterCd <- parameterCd[!is.na(parameterCd)]
-
-  url_all <- "https://www.waterqualitydata.us/Codes/public_srsnames/?mimeType=json"
-  doc <- get_nldi_sources(url_all)
-  retval <- doc[["pcodes"]]
-
-  if (all(tolower(parameterCd) != "all")) {
-    retval <- retval[retval$parm_cd %in% parameterCd, ]
-  }
-
-  attr(retval, "url") <- url_all
-
-  if (any(parameterCd != "all")) {
-    if (nrow(retval) != length(unique(parameterCd))) {
-      badPcode <- parameterCd[!(parameterCd %in% retval$parm_cd)]
-      warning(
-        "The following pCodes seem mistyped, and no information was returned: ",
-        paste(badPcode, collapse = ", ")
-      )
-    }
-  }
-
-  if (nrow(retval) != sum(is.na(parameterCd.orig))) {
-    na.params <- data.frame(matrix(
-      ncol = ncol(retval),
-      nrow = sum(is.na(parameterCd.orig))
-    ))
-    names(na.params) <- names(retval)
-    retval <- rbind(retval, na.params)
-  }
-
-  # order by parameterCd.orig
-  if (!isTRUE(parameterCd.orig == "all")) {
-    retval <- retval[match(parameterCd.orig, retval$parm_cd), ]
-    retval$parm_cd <- parameterCd.orig
-  }
-
-  return(retval)
+  .Deprecated(
+    new = "check_waterdata_sample_params",
+    package = "dataRetrieval",
+    msg = "Convert to check_waterdata_sample_params('reference-list')."
+  )
 }