+++ R/read_waterdata_field_measurements.R
@@ -51,7 +51,7 @@
 #' @examplesIf is_dataRetrieval_user()
 #'
 #' \donttest{
-#' site <- "USGS-02238500"
+#' site <- "USGS-14202650"
 #' field_data_sf <- read_waterdata_field_measurements(monitoring_location_id = site)
 #'
 #' groundwater <- read_waterdata_field_measurements(monitoring_location_id = "USGS-375907091432201")